pyproject.toml
[tool.poetry]
name = "ipr"
version = "0.1.0"
description = ""
authors = ["Puria Nafisi Azizi <puria@dyne.org>"]

[tool.poetry.scripts]
start = "ipr.app:start"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
reuse = "^0.14.0"
scancode-toolkit = "^30.1.0"
sh = "^1.14.2"
SQLAlchemy = "^1.4.35"
requests = "^2.27.1"
GitPython = "^3.1.27"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
